requirements.txt
langgraph>=0.2.60
langchain-core>=0.3.20
langchain-openai>=0.2.10
openai>=1.55.0
redis>=5.0.0
chromadb>=0.5.20
tiktoken>=0.8.0
python-dotenv>=1.0.1
pandas>=2.2.0
